# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-Commerce Demo API",
    "version": "1.0.0",
    "description": "花卉電商網站 REST API"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Id"
      }
    }
  },
  "paths": {
...
    "/api/orders": {
      "post": {
        "summary": "從購物車建立訂單",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "recipientName",
                  "recipientEmail",
                  "recipientAddress",
                  "shippingMethod",
                  "isRemoteArea",
                  "isExpress"
                ],
                "properties": {
                  "recipientName": {
                    "type": "string"
                  },
                  "recipientEmail": {
                    "type": "string",
                    "format": "email"
                  },
                  "recipientAddress": {
                    "type": "string"
                  },
                  "shippingMethod": {
                    "type": "string",
                    "enum": [
                      "home_delivery",
                      "convenience_store"
                    ]
                  },
                  "isRemoteArea": {
                    "type": "boolean"
                  },
                  "isExpress": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "訂單建立成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "order_no": {
                          "type": "string"
                        },
                        "subtotal": {
                          "type": "integer"
                        },
                        "shipping_fee": {
                          "type": "integer"
                        },
                        "shipping_method": {
                          "type": "string",
                          "enum": [
                            "home_delivery",
                            "convenience_store"
                          ]
                        },
                        "is_remote_area": {
                          "type": "boolean"
                        },
                        "is_express": {
                          "type": "boolean"
                        },
                        "total_amount": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "product_name": {
                                "type": "string"
                              },
                              "product_price": {
                                "type": "integer"
                              },
                              "quantity": {
                                "type": "integer"
                              }
                            }
                          }
                        },
                        "created_at": {
                          "type": "string"
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "購物車為空或庫存不足或收件資訊缺失"
          }
        }
      },
      "get": {
        "summary": "自己的訂單列表",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "orders": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "order_no": {
                                "type": "string"
                              },
                              "subtotal": {
                                "type": "integer"
                              },
                              "shipping_fee": {
                                "type": "integer"
                              },
                              "shipping_method": {
                                "type": "string"
                              },
                              "total_amount": {
                                "type": "integer"
                              },
                              "status": {
                                "type": "string"
                              },
                              "created_at": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "summary": "訂單詳情",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "order_no": {
                          "type": "string"
                        },
                        "recipient_name": {
                          "type": "string"
                        },
                        "recipient_email": {
                          "type": "string"
                        },
                        "recipient_address": {
                          "type": "string"
                        },
                        "subtotal": {
                          "type": "integer"
                        },
                        "shipping_fee": {
                          "type": "integer"
                        },
                        "shipping_method": {
                          "type": "string",
                          "enum": [
                            "home_delivery",
                            "convenience_store"
                          ]
                        },
                        "is_remote_area": {
                          "type": "boolean"
                        },
                        "is_express": {
                          "type": "boolean"
                        },
                        "total_amount": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "product_id": {
                                "type": "string"
                              },
                              "product_name": {
                                "type": "string"
                              },
                              "product_price": {
                                "type": "integer"
                              },
                              "quantity": {
                                "type": "integer"
                              }
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "訂單不存在"
          }
        }
      }
    },
    "/api/orders/{id}/pay": {
      "patch": {
        "summary": "模擬付款（僅 NODE_ENV=test 開放，其他環境回 404）",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "action"
                ],
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": [
                      "success",
                      "fail"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "付款狀態更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "order_no": {
                          "type": "string"
                        },
                        "total_amount": {
                          "type": "integer"
                        },
                        "status": {
                          "type": "string"
                        },
                        "created_at": {
                          "type": "string"
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "product_name": {
                                "type": "string"
                              },
                              "product_price": {
                                "type": "integer"
                              },
                              "quantity": {
                                "type": "integer"
                              }
                            }
                          }
                        }
                      }
                    },
                    "error": {
                      "type": "string",
                      "nullable": true
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "action 無效或訂單狀態不是 pending"
          },
          "404": {
            "description": "訂單不存在，或非測試環境"
          }
        }
      }
    },
    "/api/orders/{id}/check-payment": {
      "post": {
        "summary": "透過綠界 QueryTradeInfo API 查詢付款狀態",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查詢成功（已付款、尚未付款，或訂單已非 pending 時回傳目前狀態）"
          },
          "400": {
            "description": "此訂單無綠界交易編號（NO_TRADE_NO）"
          },
          "404": {
            "description": "訂單不存在"
          },
          "500": {
            "description": "綠界查詢失敗，或回應的 CheckMacValue、交易編號、金額驗證不符（ECPAY_QUERY_ERROR）"
          }
        }
      }
    },
...
  },
...
}